add support for an alternate output "style" that uses whitespace instead of semicolons
The following:

if [[ -n "$myvar" ]]; then echo 1; fi

would be formatted as:

if [[ -n "$myvar" ]]
then
    echo 1
fi

when using '-style spacey', and likewise for similar constructs
("for ...; do ... ; done", etc.) that require a ';' if there
is no line break.

The '-debug' flag is temporary to aid in debugging while making changes
for this 'spacey' style and will be removed later.